Move `SharedSessionManager` directory into `SettingsManager`
