+++ src/components/SettingsManager/SharedSessionManager/SharedSessionManager.tsx
