+++ src/components/SettingsManager/SharedSessionManager/index.tsx
